Add google benchmark for BF + add the make benchmark option.
