+++ tests/unit/CMakeLists.txt
@@ -56,6 +56,3 @@ include(GoogleTest)
 gtest_discover_tests(test_hnswlib)
 gtest_discover_tests(test_bruteforce)
 
-find_package(benchmark REQUIRED)
-add_executable(bf_benchmark bf_benchmark.cpp)
-target_link_libraries(bf_benchmark VectorSimilarity)
